Fixed temporal logging so that it outputs JSON logs and added activity and workflow telemetry data to datadog
