+++ services/archetypes/worker/src/workflowInterceptors/index.ts
@@ -1,0 +1,7 @@
+import { WorkflowInterceptorsFactory } from '@temporalio/workflow'
+
+import { WorkflowMonitoringInterceptor } from '../interceptors'
+
+export const interceptors: WorkflowInterceptorsFactory = () => ({
+  inbound: [new WorkflowMonitoringInterceptor()],
+})
