+++ services/libs/logging/src/types.ts
@@ -1,3 +1,4 @@
 import * as Bunyan from 'bunyan'
 
 export type Logger = Bunyan
+export type LogLevel = Bunyan.LogLevel
